libvirt: Add virDomainSetVcpuDirtyLimit API
Introduce virDomainSetVcpuDirtyLimit API to set or cancel the
dirty page rate upper limit.

The API will throttle the virtual CPU as needed to keep their dirty
page rate within the limit set by @rate. Since it just throttles the
virtual CPU, which dirties memory, read processes in the guest OS
aren't penalized. This could, in some scenes, be used to provide
quality-of-service in the aspect of the memory workload for virtual
CPUs.

Signed-off-by: Hyman Huang(黄勇) <yong.huang@smartx.com>