qemu_command: Check if alloc queues automatically is allowed
For virtio-blk device, Qemu provide the capability to alloc
queues 1:1 with virtul CPU by default, but it could be
incompatible with source VM in the case of live migration
if the source hypervisor does not support auto-num-queues
capability, which results in the migration failure due to
the different VMState when destination load the virtio-blk
device.

To fix it, introduce disable_auto_num_queues to check if
auto-num-queues should be enabled when launching VM for
for destination hypervisor.

Signed-off-by: Hyman Huang(黄勇) <yong.huang@smartx.com>